+++ docs/CHANGELOG.md
@@ -8,7 +8,10 @@ example, in `(String s) -> {...}` the type of `s` is `@NonNull String`.
 
 **Implementation details:**
 
-Renamed `AnnotatedTypeFactory.prepareMethodForWriting()` to `wpiPrepareMethodForWriting()`.
+Renamings in `AnnotatedTypeFactory`:
+ * `prepareCompilationUnitForWriting()` => `wpiPrepareCompilationUnitForWriting()`
+ * `prepareClassForWriting()` => `wpiPrepareClassForWriting()`
+ * `prepareMethodForWriting()` => `wpiPrepareMethodForWriting()`
 
 **Closed issues:**